Changes some small layout stuff
Adds an excerpt to the blog overview page. This is excerpt is taken from
the post front matter. If you need multiline support for the excerpt,
use quotes. That seems to work.

Changes the slab css class height property to 1px instead of 2px, on the blog overview
page not all lines where the same size, putting the height to 1px seems
to fix it.

Adds gert to the team page, just to make sure that a 4th team member is
put on the next line. Also removed the title from the members. We're all
CEO's :P

Puts services on the front page and removes them from the approach page.